Experimental structures, experimental structure initializers, and a new memory deallocation method.
The String data type no longer exists, since it is an abstraction made in LLVM IR and not directly in the language. Now there is only its stack version, which is "str," which knows the exact size at compile time and behaves like a pointer.

Now there is the pointer type, which is only reserved for very internal abstractions of the language. But they are candidates for generic types (T).